fix a few errors found by findbugs
git-svn-id: svn+ssh://svn.code.sf.net/p/jedit/svn/jEdit/trunk@5053 6b1eeb88-9816-0410-afa2-b43733a0f04e

+++ org/gjt/sp/jedit/BufferHistory.java
@@ -171,10 +171,13 @@ public class BufferHistory
 
 		String lineSep = System.getProperty("line.separator");
 
+		boolean ok = false;
+
+		BufferedWriter out = null;
+
 		try
 		{
-			BufferedWriter out = new BufferedWriter(
-				new FileWriter(file1));
+			out = new BufferedWriter(new FileWriter(file1));
 
 			out.write("<?xml version=\"1.0\"?>");
 			out.write(lineSep);
@@ -227,14 +230,30 @@ public class BufferHistory
 
 			out.close();
 
-			/* to avoid data loss, only do this if the above
-			 * completed successfully */
-			file2.delete();
-			file1.renameTo(file2);
+			ok = true;
 		}
 		catch(Exception e)
 		{
 			Log.log(Log.ERROR,BufferHistory.class,e);
+		}
+		finally
+		{
+			try
+			{
+				if(out != null)
+					out.close();
+			}
+			catch(IOException e)
+			{
+			}
+		}
+
+		if(ok)
+		{
+			/* to avoid data loss, only do this if the above
+			 * completed successfully */
+			file2.delete();
+			file1.renameTo(file2);
 		}
 
 		recentModTime = file2.lastModified();
